Resolved linting issues, renamed ChildBenefits app/component to ChildBenefitsClaim @
